+++ NEWS.md
@@ -8,6 +8,8 @@ NEW FEATURES :
 * `readClusterSTDesc()` read new clusters parameters (parameters names are now sorted)
 * `readInputTS()` read new optional time Series (5 time series)
 * `readBindingConstraints()` : has a new parameter `'constraint_names'` so the user can read only the binding constraints he wants (optimization)
+* `getThematicTrimming()` to read sub section "variables selection" of file `generaldata.ini`  
+  - Use new referentials, twice for file system and one for API
 
 BUGFIXES :
 
@@ -48,6 +50,7 @@ NEW FEATURES:
 
 * `setSimulationPathAPI()` : reads and returns the new converted study version format (ex : 9.0 => 900)
 * `setSimulationPath()` / `setSimulationPathAPI()` have a new parameter `'verbose'` (default to `FALSE`) to manage diagnostic messages
+
 
 BUGFIXES :